feat: add field Priority to Problem, seperated Full name at User

+++ api/src/API/Controllers/UsersController.cs
@@ -76,7 +76,8 @@ public class UsersController(ISender sender, IUserQueries userQueries, IIdentity
         {
             Email = createUserDto.Email,
             Password = createUserDto.Password,
-            FullName = createUserDto.FullName,
+            Name = createUserDto.Name,
+            Surname = createUserDto.Surname,
             RoleId = createUserDto.RoleId
         };
 
@@ -98,21 +99,35 @@ public class UsersController(ISender sender, IUserQueries userQueries, IIdentity
             () => NotFound());
     }
     
-    [Authorize(Roles = RoleNames.Admin)]
+    [Authorize(Roles = $"{RoleNames.Admin},{RoleNames.User}")]
     [HttpDelete("delete/{userId:guid}")]
     public async Task<ActionResult<UserDto>>
         Delete([FromRoute] Guid userId, CancellationToken cancellationToken)
     {
-        var input = new DeleteUserCommand()
-        {
-            UserId = userId
-        };
+        var currentUserIdOption = identityService.GetUserId();
+        
+        return await currentUserIdOption.Match(
+            async currentUserId =>
+            {
+                var isAdmin = identityService.IsUserInRole(RoleNames.Admin);
+                if (!isAdmin && currentUserId.Value != userId)
+                {
+                    return Forbid();
+                }
+                
+                var input = new DeleteUserCommand()
+                {
+                    UserId = userId
+                };
 
-        var result = await sender.Send(input, cancellationToken);
+                var result = await sender.Send(input, cancellationToken);
 
-        return result.Match<ActionResult<UserDto>>(
-            c => UserDto.FromDomainModel(c, imageService.GetImageUrl),
-            e => e.ToObjectResult());
+                return result.Match<ActionResult<UserDto>>(
+                    c => UserDto.FromDomainModel(c, imageService.GetImageUrl),
+                    e => e.ToObjectResult());
+            },
+            () => Task.FromResult<ActionResult<UserDto>>(Unauthorized())
+        );
     }
     
     [Authorize(Roles = RoleNames.Admin)]
@@ -183,8 +198,31 @@ public class UsersController(ISender sender, IUserQueries userQueries, IIdentity
         var input = new UpdateUserCommand
         {
             UserId = userId,
-            UserName = user.UserName,
-            Email = user.Email
+            Email = user.Email,
+            Name = user.Name,
+            Surname = user.Surname,
+            PhoneNumber = user.PhoneNumber
+        };
+
+        var result = await sender.Send(input, cancellationToken);
+
+        return result.Match<ActionResult<UserDto>>(
+            u => UserDto.FromDomainModel(u, imageService.GetImageUrl),
+            e => e.ToObjectResult());
+    }
+
+    [Authorize(Roles = $"{RoleNames.Admin},{RoleNames.User}")]
+    [HttpPut("change-password/{userId:guid}")]
+    public async Task<ActionResult<UserDto>> ChangePassword(
+        [FromRoute] Guid userId,
+        [FromBody] ChangePasswordDto request,
+        CancellationToken cancellationToken)
+    {
+        var input = new ChangePasswordCommand
+        {
+            UserId = userId,
+            CurrentPassword = request.CurrentPassword,
+            NewPassword = request.NewPassword
         };
 
         var result = await sender.Send(input, cancellationToken);
